requirements.txt
openai-agents
fastapi
uvicorn[standard]
websockets
python-dotenvgoogle-api-python-client
google-auth-oauthlib
pymongo
